move Lattice to constant_folding
