+++ libfirm-rs/src/tarval.rs
@@ -119,73 +119,3 @@ impl Add for Tarval {
 }
 
 // TODO other ops for Tarval
-
-#[derive(Clone, Copy, From, Into)]
-pub struct Lattice(Tarval);
-
-// TODO impl TryFrom<Node>
-pub enum Binop {
-    Add,
-    Phi,
-}
-
-impl Lattice {
-    pub fn binop(lhs: Tarval, rhs: Tarval, op: Binop) -> Tarval {
-        let unknown = Tarval::unknown().mode();
-        let bad = Tarval::bad().mode();
-        if lhs.mode() == unknown || rhs.mode() == unknown {
-            return Tarval::unknown();
-        }
-        if lhs.mode() == bad || rhs.mode() == bad {
-            return Tarval::bad();
-        }
-        assert_eq_modes!(lhs, rhs);
-        match op {
-            Binop::Add => lhs + rhs,
-            Binop::Phi => {
-                if lhs == rhs {
-                    lhs
-                } else {
-                    Tarval::bad()
-                }
-            }
-        }
-    }
-    pub fn phi<I: Iterator<Item = Tarval>>(mut vals: I) -> Tarval {
-        let first = vals
-            .next()
-            .expect("phi must have at least one predecessor (two in fact, but we only need one)");
-        vals.fold(first, |r, v| Lattice::binop(r, v, Binop::Phi))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn lattice_binop_modes_like_lecture_slides() {
-        init();
-        let uk = Tarval::unknown();
-        let bad = Tarval::bad();
-        let i = Tarval::mj_int(23);
-        let j = Tarval::mj_int(42);
-        let x = Lattice::binop(j, i, Binop::Add);
-
-        assert_eq_modes!(x, i);
-        assert_eq_modes!(x, j);
-
-        let x = Lattice::binop(uk, uk, Binop::Add);
-        assert_eq_modes!(x, uk);
-
-        let x = Lattice::binop(bad, bad, Binop::Add);
-        assert_eq_modes!(x, bad);
-
-        // mixed
-        let x = Lattice::binop(i, bad, Binop::Add);
-        assert_eq_modes!(x, bad);
-
-        let x = Lattice::binop(i, uk, Binop::Add);
-        assert_eq_modes!(x, uk);
-    }
-}
